+++ src/osx_main.mm
@@ -4,7 +4,47 @@
 static float WIDTH  = 1024;
 static float HEIGHT = 768;
 
-static bool Running = true;
+static bool running = true;
+static uint8_t *buffer;
+static NSBitmapImageRep *imageRep;
+
+
+static void resizeSection(int width, int height) {
+  if (imageRep) {
+    free(buffer);
+    [imageRep release];
+  }
+
+  // if (!deviceContext) {
+  //   deviceContext = ;
+  // }
+
+  int bytesPerPixel = 4;
+  int pitch = bytesPerPixel * width;
+  buffer = (uint8_t *)malloc(pitch * height);
+
+  imageRep = [[NSBitmapImageRep alloc]
+    initWithBitmapDataPlanes:&buffer
+                  pixelsWide:width
+                  pixelsHigh:height
+               bitsPerSample:8
+             samplesPerPixel:4
+                    hasAlpha:YES
+                    isPlanar:NO
+              colorSpaceName:NSDeviceRGBColorSpace
+                bitmapFormat:NSBitmapFormatThirtyTwoBitBigEndian
+                 bytesPerRow:pitch
+                bitsPerPixel:32];
+
+}
+
+static void updateWindow(NSWindow *window, int width, int height) {
+  @autoreleasepool {
+    NSImage *image = [[[NSImage alloc] initWithSize:NSMakeSize(width, height)] autorelease];
+    [image addRepresentation:imageRep];
+    window.contentView.layer.contents = image;
+  }
+}
 
 @interface WindowDelegate: NSObject<NSWindowDelegate>;
 @end
@@ -12,20 +52,18 @@ static bool Running = true;
 @implementation WindowDelegate
 
 - (void)windowWillClose:(NSNotification *)notification {
-  Running = false;
+  running = false;
 }
 
 - (void)windowDidResize:(NSNotification *)notification {
 }
 
-- (NSSize)windowWillResize:(NSWindow *)sender
-                    toSize:(NSSize)frameSize
-{
-  static bool isWhite = true;
+- (NSSize)windowWillResize:(NSWindow *)sender toSize:(NSSize)frameSize {
+  resizeSection(frameSize.width, frameSize.height);
 
-  sender.backgroundColor = isWhite ? NSColor.blackColor : NSColor.whiteColor;;
-
-  isWhite = !isWhite;
+  // static bool isWhite = true;
+  // sender.backgroundColor = isWhite ? NSColor.blackColor : NSColor.whiteColor;;
+  // isWhite = !isWhite;
 
   printf("size is %d - %d\n", (int)frameSize.width, (int)frameSize.height);
   return frameSize;
@@ -55,9 +93,8 @@ int main(int argc, const char *argv[]) {
                                                    defer:YES];
 
   window.title = @"Cocoa";
-  window.backgroundColor = NSColor.whiteColor;
+  window.backgroundColor = NSColor.blackColor;
   [window makeKeyAndOrderFront:nil];
-
 
   // [NSApp activateIgnoringOtherApps:YES];
 
@@ -67,7 +104,7 @@ int main(int argc, const char *argv[]) {
 
   printf("level is %ld\n", window.level);
 
-  while(Running) {
+  while(running) {
     NSEvent *event;
 
     do {
@@ -79,6 +116,10 @@ int main(int argc, const char *argv[]) {
         default:
           [NSApp sendEvent:event];
       }
+
+      int w = window.contentView.bounds.size.width;
+      int h = window.contentView.bounds.size.height;
+      updateWindow(window, w, h);
 
     } while(event != nil);